Working blendgamer for both 2.80 and 2.79
