+++ tools/blendgamer/src/colormap.py
