+++ tools/blendgamer/src/surfacemesh_ops.py
@@ -119,31 +119,33 @@ class GAMER_OT_fill_holes(bpy.types.Operator):
 #             return {'CANCELLED'}
 
 class SurfaceMeshImprovementProperties(bpy.types.PropertyGroup):
-    dense_rate      : FloatProperty(
+    dense_rate      = FloatProperty(
         name="CD_Rate", default=1, min=0.001, max=50.0, precision=4,
         description="The rate for coarsening dense areas")
-    dense_iter      : IntProperty(
+    dense_iter      = IntProperty(
         name="CD_Iter", default=1, min=1, max=15,
         description="The number of iterations for coarsening dense areas")
-    flat_rate       : FloatProperty(
+    flat_rate       = FloatProperty(
         name="CF_Rate", default=0.016, min=0.00001, max=0.5, precision=4,
         description="The rate for coarsening flat areas")
-    flat_iter       : IntProperty(
+    flat_iter       = IntProperty(
         name="CF_Iter", default=1, min=1, max=15,
         description="The number of iterations for coarsening flat areas")
-    smooth_iter     : IntProperty(
+    smooth_iter     = IntProperty(
         name="S_Iter", default=10, min=1, max=50,
         description="The number of iterations for coarsening dense areas")
-    preserve_ridges : BoolProperty(
+    preserve_ridges = BoolProperty(
         name="Preserve ridges", default=False,
         description="Don't flip edges which lie on ridges")
-    advanced_options : BoolProperty(name="Advanced options", default=False,
+    advanced_options = BoolProperty(
+        name="Advanced options", default=False,
         description="Show additional surface mesh improvement options")
-    autocorrect_normals : BoolProperty(name="Autocorrect normals", default=True,
+    autocorrect_normals = BoolProperty(
+        name="Autocorrect normals", default=True,
         description="Auto fix inconsistent normals")
-    verbose         : BoolProperty(name="Verbose", default=False,
+    verbose         = BoolProperty(name="Verbose", default=False,
         description="Print information to console")
-    rings           : IntProperty(
+    rings           = IntProperty(
         name="LST rings", default=2, min=1, max=5,
         description="The number of neighborhood rings to consider for LST calculation"
         )
@@ -207,16 +209,6 @@ class SurfaceMeshImprovementProperties(bpy.types.PropertyGroup):
             return gamerToBlender(report, gmesh)
         return False
 
-    # def refine_mesh(self, context, report):
-    #     gmesh = blenderToGamer(report, autocorrect_normals=self.autocorrect_normals)
-    #     if gmesh:
-    #         try:
-    #             g.refine_mesh(gmesh)
-    #         except Exception as e:
-    #             report({'ERROR'}, str(e))
-    #             return False
-    #         return gamerToBlender(report, gmesh)
-    #     return False
 
 classes = [GAMER_OT_coarse_dense,
            GAMER_OT_coarse_flat,
@@ -224,3 +216,13 @@ classes = [GAMER_OT_coarse_dense,
            GAMER_OT_normal_smooth,
            GAMER_OT_fill_holes,
            SurfaceMeshImprovementProperties]
+
+def register():
+    from bpy.utils import register_class
+    for cls in classes:
+        register_class(make_annotations(cls))
+
+def unregister():
+    from bpy.utils import unregister_class
+    for cls in reversed(classes):
+        unregister_class(make_annotations(cls))
